Scale enemy health with difficulty; use long
Compute enemy health from Global.Difficulty (health = floor(pow(1.2, Global.Difficulty) + 100)) and assign that value to the enemy Health handlers. Change HealthHandler.MaximumHealth and CurrentHealth from int to long to accommodate larger scaled values. Cast the long health to int when subtracting from Global.Health in DeathArea. Remove a couple of debug GD.Print lines from HurtHandler. Adds a System using for Math.

+++ Enemy/EnemyBasic/HealthHandler.cs
@@ -6,13 +6,14 @@ namespace tdstopdownshooter.Enemy.EnemyBasic;
 [GlobalClass]
 public partial class HealthHandler : Node
 {
-    [Export] public int MaximumHealth = 100;
-    [Export] public int CurrentHealth = 100;
+    [Export] public long MaximumHealth = 100;
+    [Export] public long CurrentHealth = 100;
 
     private ProgressBar _progressBar;
     
     public override void _Ready()
     {
+        
         _progressBar = GetNode<ProgressBar>("../EnemyHpBar");
 
         if (_progressBar != null)
